+++ Controller/Page.php
@@ -34,7 +34,6 @@ final class Page extends AbstractController
 
         // If $page isn't false, then the right $id is supplied
         if ($page !== false) {
-
             $this->loadSitePlugins();
             $this->loadBreadcrumbsByPageEntity($page);
 
@@ -43,7 +42,6 @@ final class Page extends AbstractController
             ));
 
         } else {
-
             // Returning false from controller's action triggers 404 error automatically
             return false;
         }
@@ -78,22 +76,6 @@ final class Page extends AbstractController
      */
     public function homeAction()
     {
-        return $this->homePageAction();
-        // The reason it's implemented this way, is because the CMS itself can be used to manage a single landing page
-        // Landings pages mostly contain contact form at the bottom, so the request is handled within the same action
-        if ($this->request->isPost()) {
-            return $this->submitAction();
-        } else {
-        }
-    }
-
-    /**
-     * Displays home page
-     * 
-     * @return string
-     */
-    private function homePageAction()
-    {
         $pageManager = $this->getPageManager();
         $page = $pageManager->fetchDefault();
 
@@ -107,73 +89,9 @@ final class Page extends AbstractController
             ));
 
         } else {
-
             // Returning false from a controller's action triggers 404 error automatically
             return false;
         }
-    }
-
-    /**
-     * Processes a POST request that comes from home page
-     * 
-     * @return string
-     */
-    private function submitAction()
-    {
-        // Grab all POST data
-        $input = $this->request->getPost();
-        $formValidator = $this->getValidator($input);
-
-        if ($formValidator->isValid()) {
-
-            // Send a message to site owner. Assuming the following responses will be caught by AJAX handler
-            return $this->sendMessage($input) ? '1' : '0';
-
-        } else {
-            return $formValidator->getErrors();
-        }
-    }
-
-    /**
-     * Returns prepared form validator
-     * 
-     * @param array $input
-     * @return \Krystal\Validate\ValidatorChain
-     */
-    private function getValidator(array $input)
-    {
-        return $this->validatorFactory->build(array(
-            'input' => array(
-                'source' => $input,
-                'definition' => array(
-                    // Common rules for common contact forms
-                    'name' => new Pattern\Name(),
-                    'email' => new Pattern\Email(),
-                    'message' => new Pattern\Message()
-                )
-            )
-        ));
-    }
-
-    /**
-     * Sends a message from the input
-     * 
-     * @param array $input Raw input data
-     * @return boolean
-     */
-    private function sendMessage(array $input)
-    {
-        // Render the body firstly
-        $message = $this->view->renderRaw($this->moduleName, 'messages', 'message', array(
-            'input' => $input
-        ));
-
-        // Prepare a subject
-        $subject = $this->translator->translate('You have received a new message from %s', $input['name']);
-
-        // Grab mailer service
-        $mailer = $this->getService('Cms', 'mailer');
-        return $mailer->send($subject, $message, 'You have received a new message');
     }
 
     /**